+++ src/br/alura/aula/objeto/Obj2.java
@@ -1,0 +1,13 @@
+package br.alura.aula.objeto;
+
+public class Obj2 extends Heranca{
+    String obj2;
+
+    public String getObj2() {
+        return obj2;
+    }
+
+    public void setObj2(String obj2) {
+        this.obj2 = obj2;
+    }
+}
